% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-call.R
\name{cluster_call}
\alias{cluster_call}
\title{Call a function on each node of a cluster}
\usage{
cluster_call(.cl, .fun, ...)
}
\arguments{
\item{.cl}{A cluster}

\item{.fun}{Function to call. Either a string or function. If a function,
will be copied to remote nodes.}

\item{...}{Arguments to fun}
}
\value{
A list, with one element for each node.
}
\description{
Call a function on each node of a cluster
}
\section{S3}{


This is an S3 generic, and is the only method that \pkg{multidplyr} needs
in order to use a cluster.
}
\examples{
cl <- get_default_cluster()

# When given function object, it's copied to each node:
f <- function() 0
cl \%>\% cluster_call(f)

# When given a string, the function with that name is
# called on each node
cl \%>\%
  cluster_assign_each("f", list(function() 1, function() 2)) \%>\%
  cluster_call("f")
}

