% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-eval.R
\name{cluster_eval}
\alias{cluster_eval}
\alias{cluster_eval_}
\title{Evaluate arbitrary code on each node}
\usage{
cluster_eval(cluster, expr, env = globalenv())

cluster_eval_(cluster, expr, env = globalenv())
}
\arguments{
\item{cluster}{Cluster to work on}

\item{expr}{Expression to evaluate. \code{cluster_eval} uses non-standard
evaluation to automatically quote. For \code{cluster_eval_}, you must
quote it yourself.}
}
\description{
Evaluate arbitrary code on each node
}
\examples{
cl <- create_cluster(2)
cl \%>\% cluster_eval(1 + 1)
cl \%>\% cluster_eval_(quote(1 + 1))
}

