% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-library.R
\name{cluster_library}
\alias{cluster_library}
\title{Attach a library on each node.}
\usage{
cluster_library(cluster, packages)
}
\arguments{
\item{cluster}{Cluster to work on}

\item{packages}{A character vector of package names}
}
\description{
Also attached the library in the current session.
}
\examples{
cl <- create_cluster(2)
cl \%>\% cluster_eval(search())

cl \%>\%
  cluster_library("magrittr") \%>\%
  cluster_eval(search())

cl \%>\%
  cluster_library(c("magrittr", "stats")) \%>\%
  cluster_eval(search())
}

