% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-create.R
\name{create_cluster}
\alias{create_cluster}
\title{Create a new cluster with sensible defaults.}
\usage{
create_cluster(cores = NA, quiet = FALSE)
}
\arguments{
\item{cores}{Number of cores to use. If \code{NA} will uses at least two
cores, but if yo have extra it leaves one free for other stuff.}

\item{quiet}{If \code{TRUE}, don't display initialisation message.}
}
\description{
Clusters created with this function will automatically clean up after
themselves. will automatically automatically clean up after itself.
}

