% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterman.R
\name{default_cluster}
\alias{default_cluster}
\alias{get_default_cluster}
\alias{set_default_cluster}
\title{Cluster management.}
\usage{
set_default_cluster(x)

get_default_cluster()
}
\description{
For parallel operations in \code{\link{do}}, dplyr maintains a light-weight
local cluster. Cluster creation is relatively expensive (around a second)
}
\examples{
get_default_cluster()
}

