% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-affinity.R
\name{get_core_affinity}
\alias{get_core_affinity}
\title{Assign core number based on grouping vars}
\usage{
get_core_affinity(data, grouping_vars, n_cores = parallel::detectCores())
}
\arguments{
\item{data}{Dataset}

\item{grouping_vars}{Character vector of grouping variables}

\item{n_cores}{Number of cores to divide the data among. Default parallel::detectCores()}
}
\description{
Balance core load while maintaining groups together
}
\examples{
mtcars \%>\%
  get_core_affinity(c('cyl', 'am')) \%>\%
  partition(core_n) \%>\%
  group_by(cyl, am) \%>\%
  arrange(mpg) \%>\%
  distinct(gear) \%>\%
  collect() \%>\%
  ungroup() \%>\%
  select(-core_n)
}
