% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidplyr-package.R
\docType{package}
\name{multidplyr-package}
\alias{multidplyr}
\alias{multidplyr-package}
\title{multidplyr: Partitioned data frames for 'dplyr'}
\description{
A dplyr backend that partitions a data frame across
    multiple nodes in a cluster (e.g. cores on your computer) to make
    common operations faster.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadley/multidplyr}
  \item Report bugs at \url{https://github.com/hadley/multidplyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
