% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{new_cluster}
\alias{new_cluster}
\title{Create a new cluster with sensible defaults.}
\usage{
new_cluster(n = NULL)
}
\arguments{
\item{n}{Number of workers to create. If `NULL`, it will use at least
two, but no more than the number of cores on your computer - 2.}
}
\description{
Clusters created with this function will automatically clean up after
themselves.
}
\examples{
cluster <- new_cluster(2)
cluster
}
