% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard.R
\name{shard}
\alias{shard}
\title{Shard data across a cluster.}
\usage{
shard(.data, cluster = cluster_get())
}
\description{
Shard data across a cluster.
}
\examples{
s <- shard(mtcars)
s \%>\% mutate(cyl2 = 2 * cyl)
s \%>\% filter(vs == 1)
s \%>\% summarise(n())
s \%>\% select(-cyl)
}

