% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_evenly.R
\name{spread_evenly}
\alias{spread_evenly}
\title{Allocate different sized groups into partitions}
\usage{
spread_evenly(groups, m)
}
\arguments{
\item{groups}{A tibble with columns id, n, part_id.}

\item{m}{Number of partitions to allocate into.}
}
\value{
A tibble with columns id, n, part_id, spread (approximately) evenly.
}
\description{
Implements a first fit pass plus a best fit pass heuristic of the bin
packing problem. Inspired by the algorithm here:
\url{http://stackoverflow.com/questions/16588669/spread-objects-evenly-over-multiple-collections}
If the number of groups is large, the bin packing algorithm can be expensive.
So, if the ratio of the number of groups to partitions is greater than 300, I simply do the first pass.
If the ratio is greater than 1500, I randomize the groups.
}

