% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_cluster}
\alias{src_cluster}
\title{A cluster.}
\usage{
src_cluster(cluster = get_default_cluster())
}
\arguments{
\item{cluster}{Cluster to use as backend.}
}
\description{
A cluster.
}
\examples{
src_cluster()
}

