% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_call}
\alias{cluster_call}
\title{Call a function on each node of a cluster}
\usage{
cluster_call(.cl, .fun, ...)
}
\arguments{
\item{.cl}{A cluster}

\item{.fun}{Function to call. Either a string or function. If a function,
will be copied to remote nodes.}

\item{...}{Arguments to fun}
}
\value{
A list, with one element for each node.
}
\description{
Call a function on each node of a cluster
}
\section{S3}{


This is an S3 generic, and is the only method that \pkg{dparty} needs
in order to use a cluster.
}
\examples{
c <- get_default_cluster()
cluster_call(c, "ls", envir = globalenv())
}

