% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-eval.R
\name{cluster_eval}
\alias{cluster_eval}
\alias{cluster_eval_}
\title{Evaluate arbitrary code on each node}
\usage{
cluster_eval(cluster, expr)

cluster_eval_(cluster, expr)
}
\arguments{
\item{cluster}{Cluster to work on}

\item{expr}{Expression to evaluate. \code{cluster_eval} uses non-standard
evaluation to automatically quote. For \code{cluster_eval_}, you must
quote it yourself.}
}
\description{
Note that you can only evaluate in the global environment: it doesn't
seem to make much sense to evaluate elsewhere on remote nodes.
}
\examples{
cl <- get_default_cluster()
cl \%>\% cluster_eval(1 + 1)
cl \%>\% cluster_eval_(quote(1 + 1))
}

