% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-utils.R
\name{cluster_utils}
\alias{cluster_utils}
\alias{cluster_assign}
\alias{cluster_assign_each}
\alias{cluster_copy}
\alias{cluster_rm}
\alias{cluster_ls}
\alias{cluster_library}
\title{Cluster utitility functions}
\usage{
cluster_assign(cluster, ...)

cluster_assign_each(cluster, ...)

cluster_copy(cluster, names, env = caller_env())

cluster_rm(cluster, names)

cluster_ls(cluster)

cluster_library(cluster, packages)
}
\arguments{
\item{cluster}{Cluster to work on}

\item{...}{Name-value pairs}

\item{names}{Name of variables to copy.}

\item{env}{Environment in which to look for varibles to copy.}

\item{packages}{Character vector of packages to load}
}
\value{
All functions that modify the worker environment invisibly return
  the cluster so they can chain them together. The other functions return
  lists with with one element for each worker.
}
\description{
These functions provide useful helpers for performaning common operations.
}
\examples{
cl <- default_cluster()
cluster_assign(cl, a = runif(1))
cluster_call(cl, a)

# Assign different values on each cluster
cluster_assign_each(cl, b = c(1, 10))
cluster_call(cl, b)

# If you want different to compute different values on each
# worker, use `cluster_call()` directly:
cluster_call(cl, c <- runif(1))
cluster_call(cl, c)

# cluster_copy() is a useful shortcut
d <- 10
cluster_copy(cl, "d")

cluster_ls(cl)
cluster_rm(cl, letters[1:4])
cluster_ls(cl)

# Use cluster_library() to load packages
cluster_call(cl, search())
cluster_library(cl, "magrittr")
cluster_call(cl, search())
}
