% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{default_cluster}
\alias{default_cluster}
\alias{set_default_cluster}
\alias{get_default_cluster}
\title{Cluster management.}
\usage{
set_default_cluster(x)

get_default_cluster(...)
}
\arguments{
\item{x}{New cluster to use as default.}
}
\description{
This manages a default cluster for the session, creating one the first
time it is called.
}
\examples{
get_default_cluster()
}
\keyword{internal}
