% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{default_cluster}
\alias{default_cluster}
\title{Default cluster}
\usage{
default_cluster(n = 2)
}
\arguments{
\item{n}{Number of workers to use; defaults to 2 because this is the maximum
allowed by CRAN.}
}
\value{
A cached cluster of workers.
}
\description{
Setting up a cluster is relatively expensive, so it's best to use a single
cluster throughout a session. This function lazily creates a 2-worker
cluster for use in examples and test.
}
\examples{
default_cluster()
}
\keyword{internal}
