% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidplyr-package.R
\docType{package}
\name{multidplyr-package}
\alias{multidplyr}
\alias{multidplyr-package}
\title{multidplyr: A Multi-Process 'dplyr' Backend}
\description{
Partition a data frame across multiple worker processes to provide simple multicore parallelism.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://multidplyr.tidyverse.org}
  \item \url{https://github.com/tidyverse/multidplyr}
  \item Report bugs at \url{https://github.com/tidyverse/multidplyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
