% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{new_cluster}
\alias{new_cluster}
\title{Create a new cluster with sensible defaults.}
\usage{
new_cluster(n)
}
\arguments{
\item{n}{Number of workers to create. Avoid setting this higher than the
number of cores in your computer as it will degrade performance.}
}
\description{
Clusters created with this function will automatically clean up after
themselves.
}
\examples{
cluster <- new_cluster(2)
cluster
}
