% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard.R
\name{partition}
\alias{partition}
\title{Partition data across a cluster.}
\usage{
partition(.data, ..., cluster = NULL)
}
\arguments{
\item{.data}{Dataset to partition}

\item{...}{Variables to partition by. Will generally work best when you
have many more groups than nodes. If omitted, will randomly partition
rows across nodes.}

\item{cluster}{Cluster to use. If NULL, will choose the number of nodes to be
the minimum of the number of groups and the number of cores - 1.}
}
\description{
Partition data across a cluster.
}
\examples{
library(dplyr)
s <- partition(mtcars)
s \%>\% mutate(cyl2 = 2 * cyl)
s \%>\% filter(vs == 1)
s \%>\% summarise(n())
s \%>\% select(-cyl)

if (require("nycflights13")) {
planes <- partition(flights, tailnum)
summarise(planes, n())

month <- partition(flights, month)
month \%>\% group_by(day) \%>\% summarise(n())
}
}

