% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard.R
\name{partition}
\alias{partition}
\title{Partition data across a cluster.}
\usage{
partition(.data, ..., .cluster = get_default_cluster())
}
\arguments{
\item{.data, data}{Dataset to partition}

\item{...}{Variables to partition by. Will generally work best when you
have many more groups than workers. If omitted, will randomly partition
rows across workers.}

\item{.cluster}{Cluster to use.}
}
\description{
Partition data across a cluster.
}
\examples{
library(dplyr)
cl <- new_cluster(2)

mtcars2 <- partition(mtcars, .cluster = cl)
mtcars2 \%>\% mutate(cyl2 = 2 * cyl)
mtcars2 \%>\% filter(vs == 1)
mtcars2 \%>\% group_by(cyl) \%>\% summarise(n())
mtcars2 \%>\% select(-cyl)
}
