% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard.R
\name{partition}
\alias{partition}
\title{Partition data across a cluster.}
\usage{
partition(.data, cluster = get_default_cluster())
}
\arguments{
\item{.data}{Dataset to partition}

\item{cluster}{Cluster to use.}
}
\description{
Partition data across a cluster.
}
\examples{
library(dplyr)
s <- partition(mtcars)
s \%>\% mutate(cyl2 = 2 * cyl)
s \%>\% filter(vs == 1)
s \%>\% summarise(n())
s \%>\% select(-cyl)
}

