% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partydf.R
\name{party_df}
\alias{party_df}
\title{Create party_df "by hand"}
\usage{
party_df(cluster, name)
}
\arguments{
\item{cluster}{A cluster}

\item{name}{Name of data frame variable. Must exist on every worker,
be a data frame, and have the same names.}
}
\description{
Use this function to create a partitioned data frame from existing
data frames spread across a cluster
}
\examples{
# If a real example, you might spread file names across the clusters
# and read in using data.table::fread()/vroom::vroom()/qs::qread().
cl <- default_cluster()
cluster_send(cl[1], n <- 10)
cluster_send(cl[2], n <- 15)
cluster_send(cl, df <- data.frame(x = runif(n)))

df <- party_df(cl, "df")
df
}
