% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_cluster}
\alias{src_cluster}
\title{A cluster.}
\usage{
src_cluster(cluster = get_default_cluster())
}
\arguments{
\item{cluster}{Cluster to use as backend.}
}
\description{
A cluster.
}
\examples{
library(dplyr)
cl <- src_cluster()

# Copying a data frame to a src_cluster copies it once to each node.
# This isn't usually useful unless you plan on joining it to a data
# frame created with partition()
copy_to(cl, mtcars)
}
